feature/myeong

+++ src/main/java/com/allin/filmface/choiceContents/youtube/repository/YoutubeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface YoutubeRepository extends JpaRepository<Youtube, Long> {
+public interface YoutubeRepository extends JpaRepository<Youtube, Integer> {
     // Emotion의 결과와 회원 번호에 따라 Youtube 데이터를 조회합니다.
     List<Youtube> findByEmotion_EmotionResultAndMemberNo(String emotionResult, Integer memberNo);
